Add fixes for test_client.py
