+++ mlflow/fastai.py
@@ -16,8 +16,6 @@ import os
 import yaml
 import tempfile
 import shutil
-import pandas as pd
-import numpy as np
 
 from mlflow import pyfunc
 from mlflow.models import Model, ModelSignature, ModelInputExample
@@ -235,6 +233,8 @@ class _FastaiModelWrapper:
     def predict(self, dataframe):
         from fastai.tabular import TabularList
         from fastai.basic_data import DatasetType
+        import pandas as pd
+        import numpy as np
 
         test_data = TabularList.from_df(dataframe, cont_names=self.learner.data.cont_names)
         self.learner.data.add_test(test_data)
