+++ mlflow/pytorch/__init__.py
@@ -594,6 +594,8 @@ class _PyTorchWrapper(object):
 
     def predict(self, data, device="cpu"):
         import torch
+        import numpy as np
+        import pandas as pd
 
         if not isinstance(data, pd.DataFrame):
             raise TypeError("Input data should be pandas.DataFrame")
