+++ mlflow/spacy.py
@@ -12,7 +12,6 @@ spaCy (native) format
 import logging
 import os
 
-import pandas as pd
 import yaml
 
 import mlflow
@@ -245,6 +244,7 @@ class _SpacyModelWrapper:
                           expected shape is (n_rows,1 column)
         :return: dataframe with predictions
         """
+        import pandas as pd
         if len(dataframe.columns) != 1:
             raise MlflowException("Shape of input dataframe must be (n_rows, 1column)")
