+++ tests/tracking/test_model_registry_client.py
@@ -1,0 +1,97 @@
+import pytest
+from unittest import mock
+
+from mlflow.entities.model_registry import ModelVersion, ModelVersionTag
+from mlflow.exceptions import MlflowException
+from mlflow.protos.databricks_pb2 import ErrorCode, FEATURE_DISABLED
+from mlflow.tracking import MlflowClient
+from mlflow.utils.file_utils import TempDir
+
+
+@pytest.fixture
+def mock_registry_store():
+    with mock.patch("mlflow.tracking._model_registry.utils._get_store") as mock_get_store:
+        yield mock_get_store.return_value
+
+
+def test_client_registry_operations_raise_exception_with_unsupported_registry_store():
+    """
+    This test case ensures that Model Registry operations invoked on the `MlflowClient`
+    fail with an informative error message when the registry store URI refers to a
+    store that does not support Model Registry features (e.g., FileStore).
+    """
+    with TempDir() as tmp:
+        client = MlflowClient(registry_uri=tmp.path())
+        expected_failure_functions = [
+            client._get_registry_client,
+            lambda: client.create_registered_model("test"),
+            lambda: client.get_registered_model("test"),
+            lambda: client.create_model_version("test", "source", "run_id"),
+            lambda: client.get_model_version("test", 1),
+        ]
+        for func in expected_failure_functions:
+            with pytest.raises(MlflowException) as exc:
+                func()
+            assert exc.value.error_code == ErrorCode.Name(FEATURE_DISABLED)
+
+
+def test_update_registered_model(mock_registry_store):
+    """
+    Update registered model no longer supports name change.
+    """
+    expected_return_value = "some expected return value."
+    mock_registry_store.rename_registered_model.return_value = expected_return_value
+    expected_return_value_2 = "other expected return value."
+    mock_registry_store.update_registered_model.return_value = expected_return_value_2
+    res = MlflowClient(registry_uri="sqlite:///somedb.db").update_registered_model(
+        name="orig name", description="new description"
+    )
+    assert expected_return_value_2 == res
+    mock_registry_store.update_registered_model.assert_called_once_with(
+        name="orig name", description="new description"
+    )
+    mock_registry_store.rename_registered_model.assert_not_called()
+
+
+def test_create_model_version(mock_registry_store):
+    """
+    Basic test for create model version.
+    """
+    mock_registry_store.create_model_version.return_value = _default_model_version()
+    res = MlflowClient(registry_uri="sqlite:///somedb.db").create_model_version(
+        "orig name", "source", "run-id", tags={"key": "value"}, description="desc"
+    )
+    assert res == _default_model_version()
+    mock_registry_store.create_model_version.assert_called_once_with(
+        "orig name", "source", "run-id", [ModelVersionTag(key="key", value="value")], None, "desc",
+    )
+
+
+def test_update_model_version(mock_registry_store):
+    """
+    Update registered model no longer support state changes.
+    """
+    mock_registry_store.update_model_version.return_value = _default_model_version()
+    res = MlflowClient(registry_uri="sqlite:///somedb.db").update_model_version(
+        name="orig name", version="1", description="desc"
+    )
+    assert _default_model_version() == res
+    mock_registry_store.update_model_version.assert_called_once_with(
+        name="orig name", version="1", description="desc"
+    )
+    mock_registry_store.transition_model_version_stage.assert_not_called()
+
+
+def test_transition_model_version_stage(mock_registry_store):
+    name = "Model 1"
+    version = "12"
+    stage = "Production"
+    expected_result = ModelVersion(name, version, creation_timestamp=123, current_stage=stage)
+    mock_registry_store.transition_model_version_stage.return_value = expected_result
+    actual_result = MlflowClient(registry_uri="sqlite:///somedb.db").transition_model_version_stage(
+        name, version, stage
+    )
+    mock_registry_store.transition_model_version_stage.assert_called_once_with(
+        name=name, version=version, stage=stage, archive_existing_versions=False
+    )
+    assert expected_result == actual_result
